SQLite3 SLEEPdatabase class
created a class named SLEEPdatabase for initialization and access of the database for the backend of the application

The class includes initialization of tables for sleep tracking (sleep_tracker), and sleep planning (sleep_plans) feature of the application. The class also includes methods for inserting data, fetching all the data, removing a row, and updating data from the tables where each tables have their own respective methods for these functions.